Delay lead process until `wp_loaded` and only show debug info for logged in user
Lead processing delayed until `wp_loaded` because it's late enough for WordPress
core features, plugins, themes, etc. to be ready and available but still early
enough to send headers because no output has been sent.

Debug info is also only shown to logged in users because it would be far too
revealing for public users (not to mention annoying!)