Fix #39: Doesn't work with ajax submit
https://bitbucket.org/richjenks/wp-marketo-leads/issue/39/doesnt-work-with-ajax-submit

The problem was that this plugin checks we're not in the back-end before running to prevent login forms, etc. from
triggering it. However, ajax submission makes a call to the back-end which returns false and prevents submission.

By removing this check and only looking for login or registration pages, the plugin runs with ajax submissions and at
no loss, because it only runs on specified hooks anyway.